Create git-utils.test.ts for git-utils.ts
